refactor: reorganize auth options and fix build issues
