+++ src/app/estoque/page.tsx
@@ -1,0 +1,4 @@
+export default function EstoquePage() {
+  // Placeholder component to satisfy Next.js module requirements.
+  return <div />;
+}
