+++ src/types/Category.ts
@@ -1,5 +1,5 @@
 export type Category = {
   id: number;
   title: string;
-  clientId: number;
+  clientId?: number | null;
 };
